ci: Add auto-squash merge
Without this patch, PRs are not ensured to be properly formatted,
as ther person merging will need to have this little text box to
format the message before merge.

This is not a big problem, but I think it's better to have
a consistent message, and PR message text is easy to review.

This automatically ensures the ready to merge PRs go
through a bot taking the PR message and merging all its
commits into the target branch, as a single commit, by
squashing them together.